requirements.txt
asgiref==3.7.0
Django==4.2.1
django-environ==0.10.0
djangorestframework==3.14.0
Markdown==3.4.3
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.0
